# performance improvement on big limitsets