To test performance, I was running this code:
```lua
copas.limit = require 'copas.limit'
copas.autoclose = false
local q = copas.limit.new(10)

copas.addthread(function()
  local dbs = {}
  for i = 1, 10 do
    q:addthread(function(i)
      dbs[i] = pgdriver:new{socketWrapper=copas.wrap}
    end, i)
  end
  q:wait()
  print 'connected'
  for i = 1, 100000 do
    q:addthread(function()
      local db = table.remove(dbs)
      for row in db:query(sqlquery) do
        dosomething(row)
      end
      table.insert(dbs, db)
    end)
  end
  q:wait()
end)
copas.loop()
```
Which means that in practice 100000 tasks were added into the limitset at the very beginning, and thus any `table.remove(self.queue, coro)` in the limitset had O(n); implying a total O(n²) runtime.

In this PR, the queue was changed from a table to a doubly linked list + hash table, ensuring all the operations add/remove/next, have average O(1) each.

With 10k tasks the total time went from 1.3s to 0.7s in my computer, and with 100k tasks, the time improved from 50 ~ 60s to 6.6s.